Delete confirmation dialog added

+++ src/app/categories/categories.component.ts
@@ -4,6 +4,8 @@ import { Category } from '../models/category.model';
 import { CommonModule } from '@angular/common';
 import { Router, RouterModule } from '@angular/router';
 import { finalize, tap } from 'rxjs';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmationDialogComponent } from '../confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-categories',
@@ -15,7 +17,7 @@ import { finalize, tap } from 'rxjs';
 export class CategoriesComponent {
   categories: Category[] = [];
 
-  constructor(private categoryService: CategoryService, private router: Router, private cdr: ChangeDetectorRef) { }
+  constructor(private categoryService: CategoryService, private router: Router, private cdr: ChangeDetectorRef, private dialog: MatDialog) { }
 
   ngOnInit(): void {
     this.getCategories();
@@ -29,10 +31,22 @@ export class CategoriesComponent {
   }
 
   deleteCategory(categoryId: number) {
-    this.categoryService.deleteCategory(categoryId).pipe(
-      finalize(() => {
-        console.log("silme işlemi başarılı")
-        this.ngOnInit()
-      })).subscribe()
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      width: '250px',
+      data: {
+        title: 'Category Sil',
+        message: 'Category silmek istediğinize emin misiniz?'
+      }
+    });
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.categoryService.deleteCategory(categoryId).pipe(
+          finalize(() => {
+            console.log("silme işlemi başarılı")
+            this.ngOnInit()
+          })).subscribe()
+      }
+    });
+
   }
 }
